chore: Update default igra-rpc-provider branch to main
This commit changes the default branch specified for the
`igra-rpc-provider` repository within the `setup-repos.sh` script.

The default branch was previously set to `igor/fix/docker-workaround`.
It is now updated to `main`. This ensures that running the setup script
without providing specific branch arguments will use the standard main
branch for the `igra-rpc-provider`, reflecting the most current stable
or intended default state for development and testing setups.